Fix #contact scroll position
Navigating to /contact was not scrolling to the correct section because the iframe had not loaded in yet. It would hit the bottom of page and was unable to scroll further, leaving the user at the FAQ section.

By adding a placeholder with the height of the iframe, it is now able to scroll all the way to the contact section.